gtp_bpf_prog: add soft bpf program reload feature
Add a 'soft reload': load a bpf file again, and if load is successful,
update program link with the new bpf program with
bpf_link__update_program(). If load is not successful, running bpf
program is not affected.

Only bpf code is updated. Bpf Maps and whole userapp are kept.

This soft reload is currently triggered with inotify: if bpf_prog->path
is changed on filesystem, then it is reloaded.

Not (enough) tested.

Signed-off-by: Olivier Gournet <gournet.olivier@gmail.com>